requirements.txt
Flask>=3.0.0,<4.0.0
line-bot-sdk>=3.22.0,<4.0.0
openai>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
gunicorn>=21.0.0,<24.0.0
requests>=2.32.0,<3.0.0
